[CoreLib] db 컬럼 수정, 코드퍼스트 테스트

+++ CoreSystem/CoreLibrary/Database/UserEntity.cs
@@ -5,7 +5,7 @@ namespace CoreLibrary.Database;
 
 public partial class UserEntity
 {
-    public int UserUid { get; set; }
+    public long UserUid { get; set; }
 
     public int? Level { get; set; }
